Fixed Vulkan loading fault on Apple systems
Previously any application using a Vulkan renderer would crash when
attempting to load Vulkan functions on macOS. This has been tracked
down to a change in volk (which we're using to load the Vulkan
loader), and so its submodule here has been reverted to the last
working version, which is 02295c2.

Amendment: also updated the TLVK_FuncSet_t struct to re-mirror the
now-different VolkDeviceTable struct.